requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==3.0.4
Django==3.1.4
djangorestframework==3.12.2
idna==2.10
mysqlclient==2.0.2
pytz==2020.4
requests==2.25.0
sqlparse==0.4.1
urllib3==1.26.2
